moving ShoppingCart class into model_objects.py
removing unused shopping_cart.py
