+++ python/model_objects.py
@@ -24,6 +24,7 @@ class SpecialOfferType(Enum):
     TWO_FOR_AMOUNT = 3
     FIVE_FOR_AMOUNT = 4
 
+
 class Offer:
     def __init__(self, offer_type, product, argument):
         self.offer_type = offer_type
@@ -36,6 +37,31 @@ class Discount:
         self.product = product
         self.description = description
         self.discount_amount = discount_amount
+
+
+class ShoppingCart:
+
+    def __init__(self):
+        self._items = []
+        self._product_quantities = {}
+
+    @property
+    def items(self):
+        return self._items
+
+    @property
+    def product_quantities(self):
+        return self._product_quantities
+
+    def add_item(self, product):
+        self.add_item_quantity(product, 1.0)
+
+    def add_item_quantity(self, product, quantity):
+        self._items.append(ProductQuantity(product, quantity))
+        if product in self._product_quantities.keys():
+            self._product_quantities[product] = self._product_quantities[product] + quantity
+        else:
+            self._product_quantities[product] = quantity  
 
 
 class ReceiptItem:
