+++ python/tests/test_supermarket.py
@@ -2,9 +2,8 @@ import unittest
 
 from approvaltests import verify
 
-from model_objects import Product, SpecialOfferType, ProductUnit
+from model_objects import Product, SpecialOfferType, ProductUnit, ShoppingCart
 from receipt_printer import ReceiptPrinter
-from shopping_cart import ShoppingCart
 from teller import Teller
 from tests.fake_catalog import FakeCatalog
