+++ python/texttest_fixture.py
@@ -7,9 +7,8 @@ texttest -a sr -d .
 import sys,csv
 from pathlib import Path
 
-from model_objects import Product, SpecialOfferType, ProductUnit
+from model_objects import Product, SpecialOfferType, ProductUnit, ShoppingCart
 from receipt_printer import ReceiptPrinter
-from shopping_cart import ShoppingCart
 from teller import Teller
 from tests.fake_catalog import FakeCatalog
